[RESTEASY-1887] Fix JPMS issue with tracing classes in split packages
