+++ providers/json-binding/src/main/resources/META-INF/services/org.jboss.resteasy.tracing.api.RESTEasyTracingInfo
@@ -1,1 +1,1 @@
-org.jboss.resteasy.tracing.api.providers.JSONBJsonFormatRESTEasyTracingInfo
+org.jboss.resteasy.tracing.providers.jsonb.JSONBJsonFormatRESTEasyTracingInfo
